Only serialize/deserialize session attributes that are not supported by Neo4j, improving performance and making attributes human readable when querying Neo4j.
NOTE: Allowing byte arrays, which *are* supported by Neo4j, to be unserialized is problematic. This is due to then not knowing if the value is actually a byte array or a value serialized to byte array form. For now then, byte arrays are not supported by the this class and instead are serialized. This could be enhanced in the future.